5956 implement options for OptionEnum as KeyValue<string,int> pairs

+++ Dev/Warewolf.Data.Tests/Options/OptionTests.cs
@@ -195,11 +195,11 @@ namespace Warewolf.Data.Tests
             Assert.AreEqual("MyEnum", optionEnum.Name);
 
             Assert.IsNull(optionEnum.Value);
-            optionEnum.Value = MyEnum.Option2;
+            optionEnum.Value = (int)MyEnum.Option2;
             Assert.AreEqual(MyEnum.Option2, optionEnum.Value);
 
             Assert.IsNull(optionEnum.Default);
-            optionEnum.Default = MyEnum.Option1;
+            optionEnum.Default = (int)MyEnum.Option1;
             Assert.AreEqual(MyEnum.Option1, optionEnum.Default);
         }
 
@@ -211,7 +211,7 @@ namespace Warewolf.Data.Tests
             var optionEnum = new OptionEnum
             {
                 Name = "MyEnum",
-                Value = MyEnum.Option2
+                Value = (int)MyEnum.Option2
             };
 
             var cloneOptionEnum = optionEnum.Clone() as OptionEnum;
@@ -227,7 +227,7 @@ namespace Warewolf.Data.Tests
             var optionEnum = new OptionEnum
             {
                 Name = "MyEnum",
-                Value = MyEnum.Option2
+                Value = (int)MyEnum.Option2
             };
 
             var expectedValue = optionEnum.CompareTo(null);
